updated theme, separated Login/Signup, Created MainLoggedInActivity with navigation drawer and Profile Activity
