+++ app/build.gradle
@@ -39,4 +39,5 @@ dependencies {
 
 
 
+
 apply plugin: 'com.google.gms.google-services'
