+++ app/src/main/java/com/example/tessa/kyc/LoginActivity.java
@@ -24,6 +24,7 @@ import com.google.android.gms.tasks.OnCompleteListener;
 import com.google.android.gms.tasks.OnFailureListener;
 import com.google.android.gms.tasks.OnSuccessListener;
 import com.google.android.gms.tasks.Task;
+import com.google.firebase.auth.AuthResult;
 import com.google.firebase.auth.FirebaseAuth;
 import com.google.firebase.auth.FirebaseUser;
 import com.google.firebase.database.DatabaseReference;
@@ -43,25 +44,10 @@ public class LoginActivity extends BaseActivity implements
         View.OnClickListener {
 
     private FirebaseAuth mAuth;
-    private DatabaseReference mDatabase;
-    private DatabaseReference usersRef;
-    private String userID;
-
-    private TextView emailView;
-    private TextView statusView;
-    private TextView idView;
-    private EditText nameView;
-    private EditText postalCodeView;
-    private EditText identifNoView;
-    private ImageView mImageView;
-    private TextView pleaseUpload;
-    private Spinner mSpinner;
-    private Button takePhotoButton;
-    private Button verifyEmailButton;
-    private StorageReference storageRef;
-    private Bitmap imageBitmap;
 
     final String TAG = "DED";
+    private EditText mEmailField;
+    private EditText mPasswordField;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -69,208 +55,112 @@ public class LoginActivity extends BaseActivity implements
         setContentView(R.layout.activity_login);
 
         mAuth = FirebaseAuth.getInstance();
-        userID = mAuth.getCurrentUser().getUid();
-        mDatabase = FirebaseDatabase.getInstance().getReference();
-        usersRef = mDatabase.child("users").child(userID);
 
-        emailView = (TextView) findViewById(R.id.Login_Email_TextView);
-        statusView = (TextView) findViewById(R.id.Login_Status_TextView);
-        idView = (TextView) findViewById(R.id.Login_ID_TextView);
-        nameView = (EditText) findViewById(R.id.Login_FirstName_EditText);
-        postalCodeView = (EditText) findViewById(R.id.Login_PostalCode_EditText);
-        identifNoView = (EditText) findViewById(R.id.Login_IdentifNo_EditText);
-        mImageView = (ImageView) findViewById(R.id.Login_ImageView);
-        pleaseUpload = (TextView) findViewById(R.id.Login_pleaseUpload_TextView);
-        mSpinner = (Spinner) findViewById(R.id.Login_Identification_spinner);
-        takePhotoButton = (Button) findViewById(R.id.Login_TakePhoto_button);
-        verifyEmailButton = (Button) findViewById(R.id.verify_email_button);
-
-        Intent intent = getIntent();
-        emailView.setText(intent.getStringExtra("E-mail"));
-        idView.setText(intent.getStringExtra("ID"));
-        if (mAuth.getCurrentUser().isEmailVerified()) statusView.setText("[VERIFIED]");
-        else statusView.setText("[UNVERIFIED]");
-
-        // Create an ArrayAdapter using the string array and a default spinner layout
-        ArrayAdapter<CharSequence> adapter = ArrayAdapter.createFromResource(this,
-                R.array.identification_array, android.R.layout.simple_spinner_item);
-
-        // Specify the layout to use when the list of choices appears
-        adapter.setDropDownViewResource(android.R.layout.simple_spinner_dropdown_item);
-
-        // Apply the adapter to the spinner
-        mSpinner.setAdapter(adapter);
-
-        // Create a storage reference from our app
-        storageRef = FirebaseStorage.getInstance().getReference();
-
-        if (mAuth.getCurrentUser()!=null &&
-                mAuth.getCurrentUser().isEmailVerified())
-            verifyEmailButton.setEnabled(false);
+        mEmailField = (EditText) findViewById(R.id.Login_email);
+        mPasswordField = (EditText) findViewById(R.id.Login_password);
     }
 
-    private void sendEmailVerification() {
-        // Disable button
-        findViewById(R.id.verify_email_button).setEnabled(false);
+    @Override
+    public void onStart() {
+        super.onStart();
+        // Check if user is signed in (non-null) and update UI accordingly.
+        FirebaseUser currentUser = mAuth.getCurrentUser();
+        Log.i(TAG, "signed in");
+        if (currentUser!=null) updateUI(currentUser);
+        else onResume();
+    }
 
-        // Send verification email
-        // [START send_email_verification]
-        final FirebaseUser user = mAuth.getCurrentUser();
 
-        user.sendEmailVerification()
-                .addOnCompleteListener(this, new OnCompleteListener<Void>() {
+    private void signIn(String email, String password) {
+        Log.d(TAG, "signIn:" + email);
+        if (!validateForm()) {
+            return;
+        }
+
+        showProgressDialog();
+
+        // [START sign_in_with_email]
+        mAuth.signInWithEmailAndPassword(email, password)
+                .addOnCompleteListener(this, new OnCompleteListener<AuthResult>() {
                     @Override
-                    public void onComplete(@NonNull Task<Void> task) {
-                        // [START_EXCLUDE]
-                        // Re-enable button
-                        findViewById(R.id.verify_email_button).setEnabled(true);
-
+                    public void onComplete(@NonNull Task<AuthResult> task) {
                         if (task.isSuccessful()) {
-                            Toast.makeText(LoginActivity.this,
-                                    "Verification email sent to " + user.getEmail(),
-                                    Toast.LENGTH_SHORT).show();
+                            // Sign in success, update UI with the signed-in user's information
+                            Log.d(TAG, "signInWithEmail:success");
+                            FirebaseUser user = mAuth.getCurrentUser();
+                            updateUI(user);
+
                         } else {
-                            Log.e(TAG, "sendEmailVerification", task.getException());
-                            Toast.makeText(LoginActivity.this,
-                                    "Failed to send verification email.",
+                            // If sign in fails, display a message to the user.
+                            Log.w(TAG, "signInWithEmail:failure", task.getException());
+                            Toast.makeText(LoginActivity.this, "Authentication failed.",
                                     Toast.LENGTH_SHORT).show();
+                            updateUI(null);
                         }
+
+                        // [START_EXCLUDE]
+                        if (!task.isSuccessful()) {
+//                            mStatusTextView.setText(R.string.auth_failed);
+//                            mStatusTextView.setVisibility(View.VISIBLE);
+                        }
+                        hideProgressDialog();
                         // [END_EXCLUDE]
                     }
                 });
-        // [END send_email_verification]
+        // [END sign_in_with_email]
     }
 
-    @Override
-    public void onClick(View v) {
-        int i = v.getId();
-        if (i == R.id.verify_email_button) {
-            sendEmailVerification();
-        }
-        if (i == R.id.sign_out_button) {
-            mAuth.signOut();
-            Intent intent = new Intent(this, SignUpActivity.class);
+    private void signOut() {
+        mAuth.signOut();
+        updateUI(null);
+    }
+
+
+    private void updateUI(FirebaseUser user) {
+        hideProgressDialog();
+        if (user != null) {
+            Intent intent = new Intent(this, MainLoggedInActivity.class);
+            intent.putExtra("E-mail", user.getEmail());
+            intent.putExtra("ID", user.getUid());
             startActivity(intent);
+        } else {
+            findViewById(R.id.email_password_buttons).setVisibility(View.VISIBLE);
+            findViewById(R.id.SignUp_email_password_fields).setVisibility(View.VISIBLE);
         }
-        if (i == R.id.Login_TakePhoto_button) {
-            dispatchTakePictureIntent();
-            takePhotoButton.setText("Retake Photo");
-        }
-        if (i == R.id.Login_UploadImage_button) {
-            if (validateImageUpload()) uploadImagetoFirebase();
-            else validateImageUpload();
-        }
-        if (i == R.id.Login_Submit_button) {
-            if (validateForm() && validateImageUpload()) {
-                String fn = nameView.getText().toString();
-                String pc = postalCodeView.getText().toString();
-                String id = identifNoView.getText().toString();
-                writeNewUser(fn, pc, id);
-                Toast.makeText(LoginActivity.this, "Submission Successful, Please wait for your physical token to be delivered", Toast.LENGTH_SHORT).show();
-                mAuth.signOut();
-                Intent intent = new Intent(this, SignUpActivity.class);
-                startActivity(intent);
-            } else {
-                validateForm();
-                validateImageUpload();
-            }
-        }
-    }
-
-    static final int REQUEST_IMAGE_CAPTURE = 1;
-
-    private void dispatchTakePictureIntent() {
-        Intent takePictureIntent = new Intent(MediaStore.ACTION_IMAGE_CAPTURE);
-        if (takePictureIntent.resolveActivity(getPackageManager()) != null) {
-            startActivityForResult(takePictureIntent, REQUEST_IMAGE_CAPTURE);
-        }
-    }
-
-    @Override
-    protected void onActivityResult(int requestCode, int resultCode, Intent data) {
-        if (requestCode == REQUEST_IMAGE_CAPTURE && resultCode == RESULT_OK) {
-            Bundle extras = data.getExtras();
-            imageBitmap = (Bitmap) extras.get("data");
-            mImageView.setImageBitmap(imageBitmap);
-            mImageView.setVisibility(View.VISIBLE);
-        }
-    }
-
-    private void uploadImagetoFirebase(){
-
-        // CALL THIS METHOD TO GET THE URI FROM THE BITMAP
-        Uri tempUri = getImageUri(getApplicationContext(), imageBitmap);
-
-        // CALL THIS METHOD TO GET THE ACTUAL PATH
-        File finalFile = new File(getRealPathFromURI(tempUri));
-
-        Uri file = Uri.fromFile(finalFile);
-        StorageReference identifImage = storageRef.child("images/"+userID+".jpg");
-
-        identifImage.putFile(file)
-                .addOnSuccessListener(new OnSuccessListener<UploadTask.TaskSnapshot>() {
-                    @Override
-                    public void onSuccess(UploadTask.TaskSnapshot taskSnapshot) {
-                        //Get a URL to the uploaded content
-//                      Uri downloadUrl = taskSnapshot.getDownloadUrl();
-                        Toast.makeText(getApplicationContext(), "Image Upload Successful", Toast.LENGTH_SHORT);
-                    }
-                })
-                .addOnFailureListener(new OnFailureListener() {
-                    @Override
-                    public void onFailure(@NonNull Exception exception) {
-                        // Handle unsuccessful uploads
-//                        Toast.makeText(getActivity(), "Image Upload failed",
-//                                Toast.LENGTH_LONG).show();
-                    }
-                });
-    }
-
-    public Uri getImageUri(Context inContext, Bitmap inImage) {
-        ByteArrayOutputStream bytes = new ByteArrayOutputStream();
-        inImage.compress(Bitmap.CompressFormat.JPEG, 100, bytes);
-        String path = MediaStore.Images.Media.insertImage(inContext.getContentResolver(), inImage, "Title", null);
-        return Uri.parse(path);
-    }
-
-    public String getRealPathFromURI(Uri uri) {
-        Cursor cursor = getContentResolver().query(uri, null, null, null, null);
-        cursor.moveToFirst();
-        int idx = cursor.getColumnIndex(MediaStore.Images.ImageColumns.DATA);
-        return cursor.getString(idx);
-    }
-
-    private boolean validateImageUpload() {
-        boolean valid = true;
-
-        if (mImageView.getVisibility()==View.GONE) {
-            pleaseUpload.setText("Please upload an image of you holding your identification documents for verification");
-            pleaseUpload.setVisibility(View.VISIBLE);
-            valid = false;
-        } else pleaseUpload.setVisibility(View.GONE);
-
-        return valid;
     }
 
     private boolean validateForm() {
         boolean valid = true;
 
-        TextView mfirstName = (TextView) findViewById(R.id.Login_FirstName_EditText);
-        String firstName = mfirstName.getText().toString();
-        if (TextUtils.isEmpty(firstName)) {
-            mfirstName.setError("Required.");
+        String email = mEmailField.getText().toString();
+        if (TextUtils.isEmpty(email)) {
+            mEmailField.setError("Required.");
             valid = false;
         } else {
-            mfirstName.setError(null);
+            mEmailField.setError(null);
+        }
+
+        String password = mPasswordField.getText().toString();
+        if (TextUtils.isEmpty(password)) {
+            mPasswordField.setError("Required.");
+            valid = false;
+        } else {
+            mPasswordField.setError(null);
         }
 
         return valid;
     }
 
-    private void writeNewUser(String fullName, String postalCode, String identifNo) {
-        FirebaseUser user =  mAuth.getCurrentUser();
-        String userId = user.getUid();
-        User nUser = new User(fullName, postalCode, identifNo);
-        usersRef.setValue(nUser);
+
+    @Override
+    public void onClick(View v) {
+        int i = v.getId();
+        if (i == R.id.email_sign_in_button) {
+            signIn(mEmailField.getText().toString(), mPasswordField.getText().toString());
+        }
+        else if (i == R.id.Login_signUp) {
+            Intent intent = new Intent(this, SignUpActivity.class);
+            startActivity(intent);
+        }
     }
 }
