+++ app/src/main/java/com/example/tessa/kyc/SignUpActivity.java
@@ -24,8 +24,6 @@ public class SignUpActivity extends BaseActivity implements
     private FirebaseAuth mAuth;
 
     final String TAG = "DED";
-    private TextView mStatusTextView;
-    private TextView mIDTextView;
     private EditText mEmailField;
     private EditText mPasswordField;
 
@@ -36,9 +34,6 @@ public class SignUpActivity extends BaseActivity implements
         setContentView(R.layout.activity_sign_up);
 
         mAuth = FirebaseAuth.getInstance();
-
-        mStatusTextView = (TextView) findViewById(R.id.SignUp_status);
-        mIDTextView = (TextView) findViewById(R.id.SignUp_UserID);
 
         mEmailField = (EditText) findViewById(R.id.SignUp_email);
         mPasswordField = (EditText) findViewById(R.id.SignUp_password);
@@ -75,7 +70,7 @@ public class SignUpActivity extends BaseActivity implements
                         } else {
                             // If sign in fails, display a message to the user.
                             Log.w(TAG, "createUserWithEmail:failure", task.getException());
-                            Toast.makeText(SignUpActivity.this, "Authentication failed.",
+                            Toast.makeText(SignUpActivity.this, "Task failed.",
                                     Toast.LENGTH_SHORT).show();
                             updateUI(null);
                         }
@@ -88,44 +83,6 @@ public class SignUpActivity extends BaseActivity implements
         // [END create_user_with_email]
     }
 
-    private void signIn(String email, String password) {
-        Log.d(TAG, "signIn:" + email);
-        if (!validateForm()) {
-            return;
-        }
-
-        showProgressDialog();
-
-        // [START sign_in_with_email]
-        mAuth.signInWithEmailAndPassword(email, password)
-                .addOnCompleteListener(this, new OnCompleteListener<AuthResult>() {
-                    @Override
-                    public void onComplete(@NonNull Task<AuthResult> task) {
-                        if (task.isSuccessful()) {
-                            // Sign in success, update UI with the signed-in user's information
-                            Log.d(TAG, "signInWithEmail:success");
-                            FirebaseUser user = mAuth.getCurrentUser();
-                            updateUI(user);
-                        } else {
-                            // If sign in fails, display a message to the user.
-                            Log.w(TAG, "signInWithEmail:failure", task.getException());
-                            Toast.makeText(SignUpActivity.this, "Authentication failed.",
-                                    Toast.LENGTH_SHORT).show();
-                            updateUI(null);
-                        }
-
-                        // [START_EXCLUDE]
-                        if (!task.isSuccessful()) {
-                            mStatusTextView.setText(R.string.auth_failed);
-                            mStatusTextView.setVisibility(View.VISIBLE);
-                        }
-                        hideProgressDialog();
-                        // [END_EXCLUDE]
-                    }
-                });
-        // [END sign_in_with_email]
-    }
-
     private void signOut() {
         mAuth.signOut();
         updateUI(null);
@@ -135,14 +92,11 @@ public class SignUpActivity extends BaseActivity implements
     private void updateUI(FirebaseUser user) {
         hideProgressDialog();
         if (user != null) {
-            Intent intent = new Intent(this, LoginActivity.class);
+            Intent intent = new Intent(this, MainLoggedInActivity.class);
             intent.putExtra("E-mail", user.getEmail());
             intent.putExtra("ID", user.getUid());
             startActivity(intent);
         } else {
-            mStatusTextView.setText(R.string.signed_out);
-            mIDTextView.setText(null);
-
             findViewById(R.id.email_password_buttons).setVisibility(View.VISIBLE);
             findViewById(R.id.SignUp_email_password_fields).setVisibility(View.VISIBLE);
         }
@@ -176,10 +130,6 @@ public class SignUpActivity extends BaseActivity implements
         int i = v.getId();
         if (i == R.id.email_create_account_button) {
             createAccount(mEmailField.getText().toString(), mPasswordField.getText().toString());
-        } else if (i == R.id.email_sign_in_button) {
-            signIn(mEmailField.getText().toString(), mPasswordField.getText().toString());
-        } else if (i == R.id.sign_out_button) {
-            signOut();
         }
     }
