+++ app/src/main/res/drawable/side_nav_bar.xml
@@ -1,0 +1,9 @@
+<shape xmlns:android="http://schemas.android.com/apk/res/android"
+    android:shape="rectangle">
+    <gradient
+        android:angle="135"
+        android:centerColor="#BBBBBB"
+        android:endColor="#BBBBBB"
+        android:startColor="#BBBBBB"
+        android:type="linear" />
+</shape>
