+++ app/src/main/res/layout/activity_sign_up.xml
@@ -19,26 +19,15 @@ android:orientation="vertical">
             android:layout_width="200dp"
             android:layout_height="200dp"
             android:contentDescription="@string/desc_firebase_lockup"
-            android:src="@drawable/kyc_full" />
+            android:src="@drawable/kyc" />
 
-        <TextView
+        <!--<TextView
             android:id="@+id/SignUp_Title"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:layout_marginBottom="@dimen/title_bottom_margin"
             android:text="@string/emailpassword_title_text"
-            android:theme="@style/ThemeOverlay.MyTitleText" />
-
-        <TextView
-            android:id="@+id/SignUp_status"
-            style="@style/ThemeOverlay.MyTextDetail"
-            android:text="@string/signed_out"
-            android:visibility="gone"/>
-
-        <TextView
-            android:id="@+id/SignUp_UserID"
-            style="@style/ThemeOverlay.MyTextDetail"
-            tools:text="Firebase User ID" />
+            android:theme="@style/ThemeOverlay.MyTitleText" />-->
 
     </LinearLayout>
 
@@ -79,28 +68,22 @@ android:orientation="vertical">
             android:id="@+id/email_password_buttons"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
-            android:layout_below="@+id/email_password_fields"
             android:orientation="vertical"
             android:paddingLeft="16dp"
             android:paddingRight="16dp">
 
-            <Button
-                android:id="@+id/email_sign_in_button"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:text="@string/sign_in"
-                android:onClick="onClick"
-                android:theme="@style/ThemeOverlay.MyDarkButton" />
-
-            <Button
-                android:id="@+id/email_create_account_button"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:text="@string/create_account"
-                android:onClick="onClick"
-                android:theme="@style/ThemeOverlay.MyDarkButton" />
         </LinearLayout>
 
     </LinearLayout>
+
+    <Button
+        android:id="@+id/email_create_account_button"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:text="@string/create_account"
+        android:layout_marginLeft="16dp"
+        android:layout_marginRight="16dp"
+        android:onClick="onClick"
+        android:theme="@style/ThemeOverlay.MyDarkButton" />
 
 </LinearLayout>
