+++ app/src/main/res/layout/content_main_logged_in.xml
@@ -1,0 +1,59 @@
+<?xml version="1.0" encoding="utf-8"?>
+<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:orientation="vertical"
+    app:layout_behavior="@string/appbar_scrolling_view_behavior"
+    tools:context="com.example.tessa.kyc.MainLoggedInActivity"
+    tools:showIn="@layout/app_bar_main_logged_in">
+
+    <TextView
+        android:id="@+id/MainLog_Welcome"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:textAlignment="center"
+        android:textSize="20dp"
+        android:text="Welcome!\n\n You are currently logged in as" />
+
+    <TextView
+        android:id="@+id/MainLog_Email_TextView"
+        android:layout_below="@+id/MainLog_Welcome"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:layout_gravity="center"
+        android:textAlignment="center"
+        android:textSize="20dp"
+        android:text="test"/>
+
+    <TextView
+        android:id="@+id/MainLog_Status_TextView"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:layout_below="@+id/MainLog_Email_TextView"
+        android:layout_gravity="center"
+        android:textAlignment="center"
+        android:textSize="20dp"
+        android:text="test"/>
+
+    <TextView
+        android:id="@+id/MainLog_ID_TextView"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:layout_below="@+id/MainLog_Status_TextView"
+        android:layout_gravity="center"
+        android:textAlignment="center"
+        android:textSize="20dp"
+        android:text="test"/>
+
+    <TextView
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:layout_below="@+id/MainLog_ID_TextView"
+        android:layout_gravity="center"
+        android:textAlignment="center"
+        android:textSize="20dp"
+        android:text="\n\n\nPhysical token is making its way to you"/>
+
+</RelativeLayout>
