+++ app/src/main/res/values-v21/styles.xml
@@ -1,0 +1,1 @@
+<resources></resources>
