+++ app/src/main/res/values/colors.xml
@@ -1,8 +1,8 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
-    <color name="colorPrimary">#058CBC</color>
-    <color name="colorPrimaryDark">#1F2449</color>
-    <color name="colorAccent">#F27B01</color>
+    <color name="colorPrimary">#bbbbbb</color>
+    <color name="colorPrimaryDark">#adadad</color>
+    <color name="colorAccent">#c70c0c</color>
 
     <color name="grey_100">#F5F5F5</color>
     <color name="grey_300">#E0E0E0</color>
